cloud-function/requirements.txt
requests==2.31.0
google-cloud-firestore==2.11.1
